Testing out some node based website scrapping tools
This is some super basic code to scrpe some of our databases. None of it works right now. Hopefully I can either find a solution to it working, or find a better tool to make it work. Or just skip this and move on to the rest of it working.